add manageBullet and new way to create bullet

+++ core/src/com/badlogic/drop/Sprites/EyeBullet.java
@@ -10,7 +10,7 @@ import com.badlogic.gdx.physics.box2d.World;
 public class EyeBullet extends Bullet {
 
 
-	public EyeBullet(World world, PlayScreen screen, int x, int y, int direction) {
+	public EyeBullet(World world, PlayScreen screen, float x, float y, int direction) {
 		super(world, screen, x, y, direction);
 		bulletDef.setUserData(this);
 		Collision.setCategoryFilter(bulletDef, Collision.MONSTERBULLET_BITS,null);
